- Fixed problem with conditional operator ?: and implicit casts of primitive types - Added warning on implicit cast from float to integer
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@1736 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/tests/test_feature/source/test_for.cpp
@@ -19,7 +19,7 @@ void Test()                                \n\
 {                                          \n\
   for (float k = 0; k< aSize; k++)         \n\
   {                                        \n\
-    myArray[MyFunction(k*aSize)] = k;      \n\
+    myArray[int(MyFunction(k*aSize))] = k; \n\
   }                                        \n\
                                            \n\
   for (int i = 0; i< aSize*aSize; i++)     \n\
